Extract ProductionFilter and DevelopmentFilter from Lavender
