+++ src/main/java/net/oneandone/lavender/filter/Lavender.java
@@ -15,22 +15,11 @@
  */
 package net.oneandone.lavender.filter;
 
-import net.oneandone.lavender.config.Properties;
-import net.oneandone.lavender.filter.processor.ProcessorFactory;
-import net.oneandone.lavender.filter.processor.RewriteEngine;
-import net.oneandone.lavender.index.Hex;
-import net.oneandone.lavender.index.Index;
-import net.oneandone.lavender.modules.DefaultModule;
-import net.oneandone.lavender.modules.Module;
-import net.oneandone.lavender.modules.Resource;
 import net.oneandone.sushi.fs.ExistsException;
 import net.oneandone.sushi.fs.Node;
 import net.oneandone.sushi.fs.World;
-import net.oneandone.sushi.fs.file.FileNode;
-import net.oneandone.sushi.xml.XmlException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.xml.sax.SAXException;
 
 import javax.management.InstanceAlreadyExistsException;
 import javax.management.MBeanRegistrationException;
@@ -41,21 +30,15 @@ import javax.servlet.Filter;
 import javax.servlet.FilterChain;
 import javax.servlet.FilterConfig;
 import javax.servlet.ServletException;
-import javax.servlet.ServletOutputStream;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.lang.management.ManagementFactory;
-import java.net.URI;
-import java.util.Enumeration;
-import java.util.List;
-import java.util.Map.Entry;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicReference;
 
 /**
- * A servlet filter that <em>lavendelizes</em> the response content in production mode.
+ * A servlet delegate that <em>lavendelizes</em> the response content in production mode.
  * In development mode, it delivers resources.
  */
 public class Lavender implements Filter, LavenderMBean {
@@ -67,32 +50,33 @@ public class Lavender implements Filter, LavenderMBean {
 
     private World world;
 
-    /** The filter configuration. */
+    private AtomicReference<Filter> delegate = new AtomicReference<>();
+
+    /** The delegate configuration. */
     protected FilterConfig filterConfig;
-
-    protected ProcessorFactory processorFactory;
-
-    protected List<Module> develModules;
 
     @Override
     public void init(FilterConfig config) throws ServletException {
         filterConfig = config;
 
         try {
-            LOG.info("Start initializing Lavender filter");
+            LOG.info("Start initializing Lavender delegate");
             // do not try ssh agent, because it crashes )with an abstract method error)
             // when jna is not in version 3.4.0. Which happens easily ...
             world = new World(false);
-            Node webapp = world.file(filterConfig.getServletContext().getRealPath(""));
 
-            if (useProductionFilter(webapp)) {
-                initProductionFilter(webapp);
+            Filter filter = null;
+            if (useProductionFilter(world)) {
+                filter = createProductionFilter();
             } else {
-                initDevelopmentFilter(world, webapp);
+                filter = createDevelopmentFilter();
             }
+            filter.init(filterConfig);
+            delegate.set(filter);
+
         } catch (ExistsException e) {
             e.printStackTrace();
-            throw new ServletException("Could not initialize Lavender filter", e);
+            throw new ServletException("Could not initialize Lavender delegate", e);
         } catch (RuntimeException se) {
             se.printStackTrace();
             LOG.error("Error in Lavender.init()", se);
@@ -102,57 +86,18 @@ public class Lavender implements Filter, LavenderMBean {
         registerMBean();
     }
 
-    private void initDevelopmentFilter(World world, Node webapp) throws ServletException {
-        long started = System.currentTimeMillis();
-        Properties properties = null;
-
-        try {
-            properties = Properties.load(Properties.file(world), false);
-            FileNode cache = properties.lockedCache(5, "lavenderServlet");
-            develModules = loadModulesFromWebapp(webapp, properties, cache);
-            processorFactory = null;
-        } catch (XmlException | IOException | SAXException e) {
-            e.printStackTrace();
-            throw new ServletException("Could not initialize Lavender development filter", e);
-        } finally {
-            unlockPropertiesCache(properties);
-        }
-        LOG.info("Lavender development filter for " + webapp + ", " + develModules.size()
-                + " resources. Init in " + (System.currentTimeMillis() - started + " ms"));
+    Filter createProductionFilter() throws ServletException {
+        return new ProductionFilter();
     }
 
-    List<Module> loadModulesFromWebapp(Node webapp, Properties properties, FileNode cache)
-            throws IOException, SAXException, XmlException {
-        return DefaultModule.fromWebapp(cache, false, webapp, properties.svnUsername, properties.svnPassword);
+    Filter createDevelopmentFilter() throws ServletException {
+        return new DevelopmentFilter();
     }
 
-    private void unlockPropertiesCache(Properties properties) {
-        if (properties != null) {
-            try {
-                properties.unlockCache();
-            } catch (IOException e) {
-                LOG.error("Could not unlock properties cache", e);
-            }
-        }
-    }
-
-    private boolean useProductionFilter(Node webapp) throws ExistsException {
-        Node indexSource = webapp.join(LAVENDER_IDX);
+    private boolean useProductionFilter(World world) throws ExistsException {
+        String contextPath = filterConfig.getServletContext().getRealPath("");
+        Node indexSource = world.file(contextPath).join(LAVENDER_IDX);
         return indexSource.exists();
-    }
-
-    private void initProductionFilter(Node webapp) throws ServletException {
-        Node indexSource = webapp.join(LAVENDER_IDX);
-        Node nodesSource = webapp.join(LAVENDER_NODES);
-        try {
-            Index index = Index.load(indexSource);
-            RewriteEngine rewriteEngine = RewriteEngine.load(index, nodesSource);
-            processorFactory = new ProcessorFactory(rewriteEngine);
-            LOG.info("Lavender prod filter");
-        } catch (IOException e) {
-            e.printStackTrace();
-            throw new ServletException("Could not initialize Lavender production filter", e);
-        }
     }
 
     private void registerMBean() {
@@ -167,199 +112,26 @@ public class Lavender implements Filter, LavenderMBean {
     }
 
     public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws IOException, ServletException {
-        if (processorFactory == null) {
-            doDevelFilter((HttpServletRequest) request, (HttpServletResponse) response, chain);
-        } else {
-            doProdFilter((HttpServletRequest) request, (HttpServletResponse) response, chain);
-        }
+        delegate.get().doFilter(request, response, chain);
     }
 
-    public void doDevelFilter(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-        if (!develIntercept(request, response)) {
-            chain.doFilter(request, response);
-            LOG.debug("[passed through: " + request.getMethod() + " " + request.getRequestURI() + ": " + response.getStatus() + "]");
-        }
-    }
-
-    public boolean develIntercept(HttpServletRequest request, HttpServletResponse response) throws IOException, ServletException {
-        String path;
-        Resource resource;
-
-        path = request.getPathInfo();
-        if (path == null || !path.startsWith("/")) {
-            return false;
-        }
-        path = path.substring(1);
-        resource = develLookup(path);
-        if (resource == null) {
-            return false;
-        }
-
-        switch (request.getMethod()) {
-            case "GET":
-                develGet(resource, request, response, true);
-                LOG.info(response.getStatus() + " GET " + path + " -> " + resource.getOrigin());
-                return true;
-            case "HEAD":
-                develGet(resource, request, response, false);
-                LOG.info(response.getStatus() + " HEAD " + path + " -> " + resource.getOrigin());
-                return true;
-            default:
-                return false;
-        }
-    }
-
-    private synchronized Resource develLookup(String resourcePath) throws IOException {
-        Resource resource;
-
-        // lookup cached stuff first
-        for (Module module : develModules) {
-            if (module.hasFiles()) {
-                resource = module.probe(resourcePath);
-                if (resource != null) {
-                    if (resource.isOutdated()) {
-                        LOG.info(resource.getOrigin() + ": outdated");
-                    } else {
-                        return resource;
-                    }
-                }
-            }
-        }
-        for (Module module : develModules) {
-            if (module.matches(resourcePath) != null) {
-                module.softInvalidate();
-                resource = module.probe(resourcePath);
-                if (resource != null) {
-                    return resource;
-                }
-            }
-        }
-        return null;
-    }
 
     public boolean getProd() {
-        return processorFactory != null;
+        return delegate.get() instanceof ProductionFilter;
     }
 
     public int getModules() {
-        return develModules != null ? -1 : develModules.size();
+        Filter filter = delegate.get();
+        if (filter instanceof DevelopmentFilter) {
+            return ((DevelopmentFilter) filter).getModulesCount();
+        }
+        return -1;
     }
 
-    public void doProdFilter(HttpServletRequest request, HttpServletResponse response, FilterChain chain)
-            throws IOException, ServletException {
-
-        StringBuffer url;
-        LavendelizeHttpServletRequest lavenderRequest;
-        LavendelizeHttpServletResponse lavenderResponse;
-
-        try {
-            url = request.getRequestURL();
-            URI requestURI = URI.create(url.toString());
-
-            // use custom request and response objects
-            lavenderRequest = new LavendelizeHttpServletRequest(request);
-            lavenderResponse = new LavendelizeHttpServletResponse(response, processorFactory,
-                    requestURI, request.getHeader("User-Agent"), request.getContextPath() + "/", Gzip.canGzip(request));
-            logRequest(url, request);
-        } catch (RuntimeException re) {
-            LOG.error("Error in Lavendelizer.doFilter()", re);
-            throw re;
-        }
-
-        // continue the request
-        // No exception handling at this point. Exceptions in processors are handled in LavendelizeOutputStream/Writer
-        chain.doFilter(lavenderRequest, lavenderResponse);
-
-        try {
-            // close the response to make sure all buffers are flushed
-            lavenderResponse.close();
-
-            logResponse(url, lavenderResponse);
-        } catch (IOException | RuntimeException e) {
-            LOG.error("Error in Lavendelizer.doFilter()", e);
-            throw e;
-        }
-    }
-
-    private void logRequest(StringBuffer url, HttpServletRequest httpRequest) {
-        if (LOG.isDebugEnabled()) {
-            LOG.debug("Entering doFilter: url=" + url);
-        }
-        if (LOG.isTraceEnabled()) {
-            LOG.trace("  Request headers: ");
-            Enumeration<String> headerNames = httpRequest.getHeaderNames();
-            while (headerNames.hasMoreElements()) {
-                String key = headerNames.nextElement();
-                String value = httpRequest.getHeader(key);
-                LOG.trace("    " + key + ": " + value);
-            }
-        }
-    }
-
-    private void logResponse(StringBuffer url, LavendelizeHttpServletResponse lavendelResponse) {
-        if (LOG.isDebugEnabled()) {
-            LOG.debug("Leaving doFilter:  url=" + url);
-        }
-        if (LOG.isTraceEnabled()) {
-            LOG.trace("  Response headers: ");
-            for (Entry<String, String> entry : lavendelResponse.getHeaders().entrySet()) {
-                LOG.trace("    " + entry.getKey() + ": " + entry.getValue());
-            }
-        }
-    }
 
     @Override
     public void destroy() {
-        if (develModules != null) {
-            for (Module module : develModules) {
-                try {
-                    module.saveCaches();
-                } catch (IOException e) {
-                    LOG.error("cannot save caches for " + module.getName() + ": " + e.getMessage(), e);
-                }
-            }
-        }
+        delegate.get().destroy();
     }
 
-    //--
-
-    public void develGet(Resource resource, HttpServletRequest request, HttpServletResponse response, boolean withBody) throws IOException {
-        String etag;
-        String contentType;
-        long contentLength;
-        ServletOutputStream out;
-        byte[] data;
-        String previousEtag;
-
-        etag = Hex.encodeString(resource.getMd5());
-        response.setDateHeader("Last-Modified", resource.getLastModified());
-        response.setHeader("ETag", etag);
-        contentType = filterConfig.getServletContext().getMimeType(resource.getPath());
-        if (contentType != null) {
-            response.setContentType(contentType);
-        }
-
-        previousEtag = request.getHeader("If-None-Match");
-        if (etag.equals(previousEtag)) {
-            LOG.debug("ETag match: returning 304 Not Modified: " + resource.getPath());
-            response.sendError(HttpServletResponse.SC_NOT_MODIFIED);
-        } else  { 		// first time through - set last modified time to now
-            data = resource.getData();
-            contentLength = data.length;
-            if (contentLength >= Integer.MAX_VALUE) {
-                throw new IOException(resource.getPath() + ": resource too big: " + contentLength);
-            }
-            if (withBody) {
-                response.setContentLength((int) contentLength);
-                out = response.getOutputStream();
-                try {
-                    response.setBufferSize(4096);
-                } catch (IllegalStateException e) {
-                    // Silent catch
-                }
-                out.write(data);
-            }
-        }
-    }
 }
